refactor(email): split email-notification-service into per-sender modules
